refactor(hooks): Remove UI prefix from hooks, move useSref to its own file
