+++ src/components/components.ts
@@ -16,3 +16,4 @@ export * from './UIRouter';
 export * from './UIView';
 export * from './UISref';
 export * from './UISrefActive';
+export * from './hooks';
