+++ src/components/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useCurrentStateAndParams';
 export * from './useFirstRenderEffect';
-export * from './useUIRouter';
+export * from './useRouter';
+export * from './useSref';
