--- src/components/hooks/useUIRouter.ts
@@ -1,11 +1,0 @@
-import { useContext } from 'react';
-import { UIRouterContext, UIRouterInstanceUndefinedError } from '../UIRouter';
-
-/** Returns the UIRouter object from React Context */
-export function useUIRouter() {
-  const router = useContext(UIRouterContext);
-  if (!router) {
-    throw new Error(UIRouterInstanceUndefinedError);
-  }
-  return router;
-}
